Make Key component appear pressed on key down
For this to work I changed the :active pseudo-class to an actual
Key_active class, which is added to the currently active Key. For this
to work I had to add state to the Keyboard component. For now it's the
same as the App's state, but I don't think I'll keep that state in the
App for a long time.