Fix input value changing to 26 on letter input
Previously, the position and ring setting inputs show 26 when a
non-number is entered. Apparently the value of a number input when a
non-number is entered is the empty string. That gets coerced to 0, then
rolls back to 26 according to the usual logic.

Now when a non-number is entered, the rotor's props no longer change.